requirements.txt
# Core
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.1
pydantic-settings==2.9.1
python-dotenv==1.1.0

# HTTP client
httpx==0.28.1

# HTML parsing
beautifulsoup4==4.13.4
readability-lxml==0.8.4.1
lxml==5.4.0

# PDF generation
playwright==1.51.0
pypdf==5.4.0

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0

# Utilities
python-multipart==0.0.20
